refactor(build): vercel serve and constants (#877)
following #866
