--- packages/waku/src/internals.ts
@@ -1,2 +1,0 @@
-export { runner as unstable_runner } from './lib/hono/runner.js';
-export { build as unstable_build } from './lib/builder/build.js';
