+++ packages/waku/src/unstable_hono.ts
@@ -1,0 +1,7 @@
+// These exports are for internal use only and subject to change without notice.
+
+export { runner } from './lib/hono/runner.js';
+
+export { Hono } from 'hono';
+export { getRequestListener } from '@hono/node-server';
+export { serveStatic } from '@hono/node-server/serve-static';
